Canonicalize plain-@testset sublibraries onto @safetestset
Several OrdinaryDiffEq.jl sublibraries still used plain `@testset` for their
top-level isolation units rather than the canonical per-unit `@safetestset`
that the root suite and most sublibraries already use. This wraps each
independent unit in `@safetestset` so it runs in its own module (isolation +
world-age safety), matching the established structure.

Converted sublibraries:
- ImplicitDiscreteSolve: 7 inline `@testset` units -> `@safetestset` (imports copied in)
- SimpleImplicitDiscreteSolve: 2 inline units -> `@safetestset`
- OrdinaryDiffEqNewmark: 6 inline units -> `@safetestset`
- OrdinaryDiffEqTaylorSeries: 10 inline units -> `@safetestset`
- OrdinaryDiffEqAMF: outer unit -> `@safetestset`
- DiffEqDevTools: 8 `@testset begin include() end` -> `@safetestset include()`
- OrdinaryDiffEqExplicitTableaus: outer unit -> `@safetestset include()` (body
  extracted to explicit_tableau_tests.jl so `const ET`/`using`/`import` run at
  module scope; an inline body would put `const` in @testset-local scope, which
  Julia rejects with "unsupported const declaration on local variable")
- OrdinaryDiffEqImplicitTableaus: same, body -> implicit_tableau_tests.jl
- OrdinaryDiffEqRosenbrockTableaus: same, body -> rosenbrock_tableau_tests.jl;
  trailing convergence unit -> `@safetestset include()`
- StochasticDiffEqLevyArea: outer unit -> `@safetestset` (imports + Random.seed!
  hoisted into the body)

Nested grouping `@testset`s inside each unit stay plain (the unit-level
@safetestset already isolates). GROUP/TEST_GROUP dispatch ladders, run paths,
QA structure, and all assertions are unchanged.

SafeTestsets added to test deps of the sublibraries that lacked it
(DiffEqDevTools, OrdinaryDiffEqExplicitTableaus, OrdinaryDiffEqImplicitTableaus,
OrdinaryDiffEqRosenbrockTableaus): [extras] uuid + [targets].test +
[compat] SafeTestsets = "0.1, 1".

Verified locally (julia +1.11, GROUP=Core): StochasticDiffEqLevyArea 60/60,
OrdinaryDiffEqImplicitTableaus 30/30, OrdinaryDiffEqExplicitTableaus 233/233,
ImplicitDiscreteSolve 30/30, SimpleImplicitDiscreteSolve 18/18 all pass.
