+++ lib/DiffEqDevTools/test/runtests.jl
@@ -1,5 +1,6 @@
 using DiffEqDevTools
 using Pkg
+using SafeTestsets
 using Test
 
 const TEST_GROUP = get(ENV, "ODEDIFFEQ_TEST_GROUP", "ALL")
@@ -12,32 +13,16 @@ end
 # Run QA tests (Aqua) — skip on pre-release Julia
 if (TEST_GROUP == "QA" || TEST_GROUP == "ALL") && isempty(VERSION.prerelease)
     activate_qa_env()
-    @time @testset "Aqua" begin
-        include("qa/qa.jl")
-    end
+    @time @safetestset "Aqua" include("qa/qa.jl")
 end
 
 if TEST_GROUP == "Core" || TEST_GROUP == "ALL"
     # write your own tests here
-    @time @testset "Benchmark Tests" begin
-        include("benchmark_tests.jl")
-    end
-    @time @testset "ODE AppxTrue Tests" begin
-        include("ode_appxtrue_tests.jl")
-    end
-    @time @testset "Analyticless Convergence Tests" begin
-        include("analyticless_convergence_tests.jl")
-    end
-    @time @testset "Analyticless Stochastic WP" begin
-        include("analyticless_stochastic_wp.jl")
-    end
-    @time @testset "Stability Region Tests" begin
-        include("stability_region_test.jl")
-    end
-    @time @testset "Plot Recipes" begin
-        include("plotrecipes_tests.jl")
-    end
-    @time @testset "Plot Recipes (Nonlinearsolve WP-diagrams)" begin
-        include("nonlinearsolve_wpdiagram_tests.jl")
-    end
+    @time @safetestset "Benchmark Tests" include("benchmark_tests.jl")
+    @time @safetestset "ODE AppxTrue Tests" include("ode_appxtrue_tests.jl")
+    @time @safetestset "Analyticless Convergence Tests" include("analyticless_convergence_tests.jl")
+    @time @safetestset "Analyticless Stochastic WP" include("analyticless_stochastic_wp.jl")
+    @time @safetestset "Stability Region Tests" include("stability_region_test.jl")
+    @time @safetestset "Plot Recipes" include("plotrecipes_tests.jl")
+    @time @safetestset "Plot Recipes (Nonlinearsolve WP-diagrams)" include("nonlinearsolve_wpdiagram_tests.jl")
 end
