+++ lib/OrdinaryDiffEqAMF/test/runtests.jl
@@ -11,7 +11,7 @@ end
 
 # Run functional tests
 if TEST_GROUP == "Core" || TEST_GROUP == "ALL"
-    @testset "OrdinaryDiffEqAMF" begin
+    @time @safetestset "OrdinaryDiffEqAMF" begin
         include("test_pollu.jl")
         include("test_fd2d.jl")
         include("test_adjoint_fd2d.jl")
