+++ lib/OrdinaryDiffEqImplicitTableaus/Project.toml
@@ -8,6 +8,7 @@ DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 
 [extras]
 DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [sources]
@@ -19,8 +20,9 @@ path = "../DiffEqDevTools"
 [compat]
 DiffEqBase = "7"
 DiffEqDevTools = "3"
+SafeTestsets = "0.1, 1"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
 [targets]
-test = ["DiffEqDevTools", "Test"]
+test = ["DiffEqDevTools", "SafeTestsets", "Test"]
