+++ lib/OrdinaryDiffEqRosenbrockTableaus/Project.toml
@@ -10,6 +10,7 @@ Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 ODEProblemLibrary = "fdc4e326-1af4-4b90-96e7-779fcce2daa5"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
@@ -19,11 +20,12 @@ Enzyme = "0.13"
 LinearAlgebra = "1.10"
 LinearSolve = "3.46"
 ODEProblemLibrary = "1"
+SafeTestsets = "0.1, 1"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
 [targets]
-test = ["ADTypes", "DiffEqDevTools", "Enzyme", "LinearAlgebra", "LinearSolve", "ODEProblemLibrary", "Test"]
+test = ["ADTypes", "DiffEqDevTools", "Enzyme", "LinearAlgebra", "LinearSolve", "ODEProblemLibrary", "SafeTestsets", "Test"]
 
 [sources.DiffEqDevTools]
 path = "../DiffEqDevTools"
